% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{array2}
\alias{array2}
\title{Fill Array with C-style row-major}
\usage{
array2(data, dims, dimnames)
}
\arguments{
\item{data}{what to fill the array with}

\item{dims}{numeric vector of dimensions}

\item{dimnames}{a list of dimnames, must be the same length as \code{dims}}
}
\description{
This fills an array as-if it were a C-style, row-major. The returned array is
still R's native F-style. Otherwise, it behaves identically to `base::array()``
}
