% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{drop_dimnames}
\alias{drop_dimnames}
\title{Drop dimnames}
\usage{
drop_dimnames(x, which_dim = NULL, keep_axis_names = FALSE)
}
\arguments{
\item{x}{an object, potentially with dimnames}

\item{.dim}{If \code{NULL} (the default) then all dimnames are dropped. If integer
vector, then dimnames only at the specified dimensions are dropped.}
}
\description{
A pipe-friendly wrapper for \code{dimnames(x) <- NULL} or (in essence)
\code{dimnames(x)[.dim] <- list(NULL)}
}
