% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_along.R
\name{reverse_dim}
\alias{reverse_dim}
\alias{reverse_rows}
\alias{reverse_cols}
\title{reverse order along a dimension}
\usage{
reverse_dim(x, .dim)

reverse_rows(x)

reverse_cols(x)
}
\arguments{
\item{x}{a dataframe, array, or vector, or anything with a \code{[} method
defined. In the case of a vector, this is equivalent to \code{rev}}
}
\description{
reverse order along a dimension
}
