% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{set_dim}
\alias{set_dim}
\title{Pipe friendly \code{dim<-}, with option to pad to necessary length}
\usage{
set_dim(x, .dim, pad = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A vector to set dimensions on}

\item{.dim}{The desired dimensions}

\item{pad}{The value to pad the vector with. \code{NULL} (the default) performs no padding.}

\item{verbose}{Whether to emit a message if padding. By default, \code{TRUE}.}
}
\value{
Object with dimensions set
}
\description{
Pipe friendly \code{dim<-}, with option to pad to necessary length
}
\examples{
set_dim(1:10, c(2, 5))
try( set_dim(1:7, c(2, 5)) ) # error by default, just like `dim<-`
     set_dim(1:7, c(2, 5), pad = 99)
}
