% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim2.R
\name{t.array}
\alias{t.array}
\title{transpose an array}
\usage{
\method{t}{array}(x)
}
\arguments{
\item{x}{an array

This reverses the dimensions of an array}
}
\description{
transpose an array
}
\examples{
x <- array(1:27, c(3,3,3))
tx <- t(x)
for (i in 1:3)
  for(j in 1:3)
    stopifnot(x[,j,i] == tx[i,j,])
}
