% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIM}
\alias{DIM}
\title{DIM}
\usage{
DIM(x)
}
\arguments{
\item{x}{an array or atomic vector}
}
\description{
\code{DIM()} is to \code{dim()} as \code{NROW()} is to \code{nrow()}. That is, it is identical to
\code{dim()} in most cases except if the input is a bare atomic vectors with no
\code{dim} attribute, in which case, the length of the vector is returned instead
of \code{NULL}.
}
