% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_dimnames}
\alias{drop_dimnames}
\alias{drop_rownames}
\alias{drop_colnames}
\title{Drop dimnames}
\usage{
drop_dimnames(x, .dim = NULL)

drop_rownames(x)

drop_colnames(x)
}
\arguments{
\item{x}{an object, potentially with dimnames}

\item{.dim}{If \code{NULL} (the default) then all dimnames are dropped. If integer
vector, then dimnames only at the specified dimensions are dropped.}
}
\description{
A pipe-friendly wrapper for \code{dimnames(x) <- NULL} or
\code{dimnames(x)[[.dim]] <- NULL}
}
