% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_on_dim}
\alias{split_on_dim}
\alias{split_on_rows}
\alias{split_along_dim}
\alias{split_along_rows}
\title{Split an array along a dimension}
\usage{
split_on_dim(X, .dim, f = dimnames(X)[[.dim]] \%||\% seq_along_dim(X, .dim),
  drop = FALSE, depth = Inf)

split_on_rows(X, f = rownames(X) \%||\% seq_along_rows(X), drop = FALSE,
  depth = Inf)

split_along_dim(X, .dim, drop = NULL, .keep_names = TRUE, depth = Inf)

split_along_rows(X, drop = NULL, .keep_names = TRUE, depth = Inf)
}
\arguments{
\item{X}{an array, or list of arrays. Atomic vectors without a dimension
attribute is treated as a 1 dimensions array (Meaning, atomic vectors
without a dim attribute are only accepted if \code{.dim} is \code{1}. Names of list
are preserved. If a list of arrays, all the arrays must have the same
length of the dimension being split.}

\item{.dim}{a scalar integer, specifying which dimension to split along}

\item{f}{Specify how to split the dimension. \describe{

\item{character, integer or factor}{passed on to \code{base::split()}. Must be
the same length as the dimention being split.}

\item{a list of vectors}{Passed on to \code{base::interaction()} then
\code{base::split()}. Each vector in the list must be the same length as the
dimention being split.}

\item{a scalar integer}{used to split into that many groups of equal size}

\item{a numeric vector where \code{all( f < 0 )} }{used to determin the
relative proportions of the group being split. \code{sum(f)} must be
\code{1}. For example \code{c(0.2, 0.2, 0.6)} will return approximatly a
20\%-20\%-60\% split.} }}

\item{drop}{passed on to \code{[}.}

\item{depth}{Scalar number, how many levels to recurse down. Set this if you
want to explicit treat a list as a vector (that is, a one-dimentional
array). (You can alternatively set dim attributes with \code{dim<-} on the list
to prevent recursion)}

\item{.keep_names}{Logical. If \code{TRUE} then if the dim being split along has
dimnames, then the returned list has those names.}
}
\value{
A list of arrays, or if a list of arrays was passed in, then a list
of lists of arrays.
}
\description{
\code{split_along_dim(X, .dim)} is equivalent to \code{split_on_dim(X, seq_along_dim(X, .dim))}
}
\examples{
X <- array(1:8, c(2,3,4))
X
split_along_dim(X, 2)
split_on_dim(X, 2, c("a", "a", "b"), drop = FALSE)
}
